+++ src/admins/__http_tests__/admin.student.list.http
@@ -1,7 +1,7 @@
 @port = {{$dotenv API_PORT}}
 @baseUrl = http://localhost:{{port}}
-@accessToken = eyJhbGciOiJQUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjIsInRva2VuVHlwZSI6ImFjY2VzcyIsInJvbGUiOiJTVVBFUl9BRE1JTiIsImlhdCI6MTc0NDI2MjU2MywiZXhwIjoxNzQ0MjYzNDYzfQ.ET0-nmV43fUHBtkll1I1imsCet4eq84GyltzFpDkSRsntTeLBCu-H-gTZfTdAkspgJLLXIr5dp16Zp3RmOBUImLGS0oESx-C4oehCVx4YM7RorFGszrV-GIRdb2CyvsPQwbvZZXf8pzzBwS6aEWrrbscM-cfCH5WA6o_fLOBzwzuWKxnUX-cx4DJ9EXxsc0zm_f7JHBgLIKmlHQMG9W180qyajrzN4AW_RhcFTsrjEXTikpsD8zjvRClkky4F4aAbUjD6Z6KomMWHjzKS7oPerPkDUgGb6He58z5RtgqpPPyT7kMnpOn661aKVI7831HOfC_M_bM1HXTy0ZpIIsqwm7XwrCzvIBV9RJxVqrqgurkKJZ22LVhzA0YBDHFTup3v3AZ4RryFoqV7gwjoHBsGC6Jsk0yZMciUBJZ3e5XiPyJmFXrqR-rUb-RAW0Mdesjz8RJD_yZMYx0dGT9rePF3nYlsEoUAzKqXA22sTcUFe-1Ej5_y3To3uA-9LX1HPSXgXERotVhW4eTUfgwCPbD2w2N34OHIHTPTedhaXWr2gkWSuZFI0O9lguQH-pyJegEbQZUdtD2OlxXJRZKb7ebX95-grsclw5WgD8Jfx80cy4R_VN-BG9zz7C1qrLQ1TxSYmzlMFX60_VHK1rG4PK1X8zGFPFmEuPGGGfuD6KJXAo
+@accessToken = eyJhbGciOiJQUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsInRva2VuVHlwZSI6ImFjY2VzcyIsInJvbGUiOiJTVVBFUl9BRE1JTiIsImlhdCI6MTc0NjQwNDgwNSwiZXhwIjoxNzQ2NDA1NzA1fQ.iqqDxNj7KNleRrg6g88jkoMu6KVaVZBf4ZlFrwzU8X9ICRXHEs4iSnAt8gqOJiuTu7YYsgIn1j-Erkdeo9wQNTbgLRSGzbD21zDbq0wJ4xeu-X9bKoMzSORPTjT5yA6Uws5wpb0_kARcq6TYysGRDLbFXPMupnQHJ1k0d9sErM2-yQRW46cM8tNDwriwb_C5Xcu_EdXQ9JNoidnTZzMuHl_PxzOugHGGKho2Chj_F5Siqip1z85vuq4C7SHpHuu-z3TgYO1WvB7RMt0bV5xsBC8fnHdk4HseXgoBM2SNU0rRJqtAAOUCfGZ78hCLzeqQmbnb_pWWCSp1ByvyaGSixr2sKxiyC2hICfM3S_t79qpXTFnrkQP4yVeDEy8EQO5JKWt0Vl6LU17f_Hd7pEvRPlaFdU4o1E9pbQVGglUKCK67ShUnjYoQWKUla4unpcHILKGzWZe_oD9qJgAqC3C8KGO9hRu2GyeGSsHqbstwVoVjBe1xQappS-uswqsi7p-_ypB-QI9OAOwqqHNSwTQHd3WmlfOnAAnO1kPqa71IzJUUb5pnkbPBNz0PVCM5oQ42QfCsYBgQGsmoHo5o6WD4JoLHTT8TrEBnx8Ak2RI69AAyZY_CIaDqzskgDQoiDE8RMX7ZPRHFKJGOA0JpW7N2bH0PmWsdbIyySqAQwWgTbig
 
-GET {{baseUrl}}/admins/users/students?status=PENDING
+GET {{baseUrl}}/admins/users/students
 Authorization: Bearer {{accessToken}}
 Content-Type: application/json
