+++ src/admins/__http_tests__/admin.student.search.http
@@ -1,6 +1,6 @@
 @port = {{$dotenv API_PORT}}
 @baseUrl = http://localhost:{{port}}
-@accessToken = 
+@accessToken = eyJhbGciOiJQUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsInRva2VuVHlwZSI6ImFjY2VzcyIsInJvbGUiOiJTVVBFUl9BRE1JTiIsImlhdCI6MTc0NjQwNDgwNSwiZXhwIjoxNzQ2NDA1NzA1fQ.iqqDxNj7KNleRrg6g88jkoMu6KVaVZBf4ZlFrwzU8X9ICRXHEs4iSnAt8gqOJiuTu7YYsgIn1j-Erkdeo9wQNTbgLRSGzbD21zDbq0wJ4xeu-X9bKoMzSORPTjT5yA6Uws5wpb0_kARcq6TYysGRDLbFXPMupnQHJ1k0d9sErM2-yQRW46cM8tNDwriwb_C5Xcu_EdXQ9JNoidnTZzMuHl_PxzOugHGGKho2Chj_F5Siqip1z85vuq4C7SHpHuu-z3TgYO1WvB7RMt0bV5xsBC8fnHdk4HseXgoBM2SNU0rRJqtAAOUCfGZ78hCLzeqQmbnb_pWWCSp1ByvyaGSixr2sKxiyC2hICfM3S_t79qpXTFnrkQP4yVeDEy8EQO5JKWt0Vl6LU17f_Hd7pEvRPlaFdU4o1E9pbQVGglUKCK67ShUnjYoQWKUla4unpcHILKGzWZe_oD9qJgAqC3C8KGO9hRu2GyeGSsHqbstwVoVjBe1xQappS-uswqsi7p-_ypB-QI9OAOwqqHNSwTQHd3WmlfOnAAnO1kPqa71IzJUUb5pnkbPBNz0PVCM5oQ42QfCsYBgQGsmoHo5o6WD4JoLHTT8TrEBnx8Ak2RI69AAyZY_CIaDqzskgDQoiDE8RMX7ZPRHFKJGOA0JpW7N2bH0PmWsdbIyySqAQwWgTbig
 
 @query = 234
